cleanup & renaming; remove static metohds; TBrake always initialize
