+++ .vimrc
@@ -1,0 +1,1 @@
+set expandtab
