--- demo/bin/example.gdextension
@@ -1,25 +1,0 @@
-[configuration]
-
-entry_symbol = "example_library_init"
-compatibility_minimum = "4.1"
-
-[libraries]
-
-macos.debug = "res://bin/libgdexample.macos.template_debug.framework"
-macos.release = "res://bin/libgdexample.macos.template_release.framework"
-ios.debug = "res://bin/libgdexample.ios.template_debug.framework"
-ios.release = "res://bin/libgdexample.ios.template_release.framework"
-windows.debug.x86_32 = "res://bin/libgdexample.windows.template_debug.x86_32.dll"
-windows.release.x86_32 = "res://bin/libgdexample.windows.template_release.x86_32.dll"
-windows.debug.x86_64 = "res://bin/windows/libmaszyna.windows.template_debug.x86_64.dll"
-windows.release.x86_64 = "res://bin/libgdexample.windows.template_release.x86_64.dll"
-linux.debug.x86_64 = "res://bin/libgdexample.linux.template_debug.x86_64.so"
-linux.release.x86_64 = "res://bin/libgdexample.linux.template_release.x86_64.so"
-linux.debug.arm64 = "res://bin/libgdexample.linux.template_debug.arm64.so"
-linux.release.arm64 = "res://bin/libgdexample.linux.template_release.arm64.so"
-linux.debug.rv64 = "res://bin/libgdexample.linux.template_debug.rv64.so"
-linux.release.rv64 = "res://bin/libgdexample.linux.template_release.rv64.so"
-android.debug.x86_64 = "res://bin/libgdexample.android.template_debug.x86_64.so"
-android.release.x86_64 = "res://bin/libgdexample.android.template_release.x86_64.so"
-android.debug.arm64 = "res://bin/libgdexample.android.template_debug.arm64.so"
-android.release.arm64 = "res://bin/libgdexample.android.template_release.arm64.so"
