+++ demo/bin/libmaszyna.gdextension
@@ -1,0 +1,26 @@
+[configuration]
+
+entry_symbol = "libmaszyna_library_init"
+compatibility_minimum = "4.1"
+reloadable = true
+
+[libraries]
+
+macos.debug = "res://bin/macos/libmaszyna.macos.template_debug.framework"
+macos.release = "res://bin/macos/libmaszyna.macos.template_release.framework"
+ios.debug = "res://bin/ios/libmaszyna.ios.template_debug.framework"
+ios.release = "res://bin/ios/libmaszyna.ios.template_release.framework"
+windows.debug.x86_32 = "res://bin/windows/libmaszyna.windows.template_debug.x86_32.dll"
+windows.release.x86_32 = "res://bin/windows/libmaszyna.windows.template_release.x86_32.dll"
+windows.debug.x86_64 = "res://bin/windows/libmaszyna.windows.template_debug.x86_64.dll"
+windows.release.x86_64 = "res://bin/windows/libmaszyna.windows.template_release.x86_64.dll"
+linux.debug.x86_64 = "res://bin/linux/libmaszyna.linux.template_debug.x86_64.so"
+linux.release.x86_64 = "res://bin/linux/libmaszyna.linux.template_release.x86_64.so"
+linux.debug.arm64 = "res://bin/linux/libmaszyna.linux.template_debug.arm64.so"
+linux.release.arm64 = "res://bin/linux/libmaszyna.linux.template_release.arm64.so"
+linux.debug.rv64 = "res://bin/linux/libmaszyna.linux.template_debug.rv64.so"
+linux.release.rv64 = "res://bin/linux/libmaszyna.linux.template_release.rv64.so"
+android.debug.x86_64 = "res://bin/android/libmaszyna.android.template_debug.x86_64.so"
+android.release.x86_64 = "res://bin/android/libmaszyna.android.template_release.x86_64.so"
+android.debug.arm64 = "res://bin/android/libmaszyna.android.template_debug.arm64.so"
+android.release.arm64 = "res://bin/android/libmaszyna.android.template_release.arm64.so"
