+++ demo/node_2d.tscn
@@ -4,3 +4,5 @@
 
 [node name="Node2D" type="Node2D"]
 script = ExtResource("1_c5xuv")
+
+[node name="MaszynaTBrake" type="MaszynaTBrake" parent="."]
