+++ demo/project.godot
@@ -10,7 +10,7 @@ config_version=5
 
 [application]
 
-config/name="godot cpp template"
+config/name="LibMaszyna Demo"
 run/main_scene="res://node_2d.tscn"
 config/features=PackedStringArray("4.3", "Forward Plus")
 config/icon="res://icon.svg"
