--- doc_classes/Brake.xml
@@ -1,92 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<class name="Brake" inherits="Node" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="https://raw.githubusercontent.com/godotengine/godot/master/doc/class.xsd">
-	<brief_description>
-	</brief_description>
-	<description>
-		Class handling brake system with brake types and parameters
-	</description>
-	<tutorials>
-	</tutorials>
-	<methods>
-		<method name="get_bcf" qualifiers="static">
-			<return type="float" />
-			<description>
-			</description>
-		</method>
-		<method name="get_bcp" qualifiers="static">
-			<return type="float" />
-			<description>
-				Brake cylinder pressure
-			</description>
-		</method>
-		<method name="get_brp" qualifiers="static">
-			<return type="float" />
-			<description>
-				Auxiliary tank pressure
-			</description>
-		</method>
-		<method name="get_crp" qualifiers="static">
-			<return type="float" />
-			<description>
-				Control tank pressure
-			</description>
-		</method>
-		<method name="get_edbcp" qualifiers="static">
-			<return type="float" />
-			<description>
-				Electrodynamic brake control pressure
-			</description>
-		</method>
-		<method name="get_fc" qualifiers="static">
-			<return type="float" />
-			<param index="0" name="_unnamed_arg0" type="float" />
-			<param index="1" name="_unnamed_arg1" type="float" />
-			<description>
-				Brake friction coefficient
-			</description>
-		</method>
-		<method name="get_hpflow" qualifiers="static">
-			<return type="float" />
-			<param index="0" name="_unnamed_arg0" type="float" />
-			<param index="1" name="_unnamed_arg1" type="float" />
-			<description>
-				Flow from the supply line
-			</description>
-		</method>
-		<method name="get_pf" qualifiers="static">
-			<return type="float" />
-			<param index="0" name="_unnamed_arg0" type="float" />
-			<param index="1" name="_unnamed_arg1" type="float" />
-			<param index="2" name="_unnamed_arg2" type="float" />
-			<description>
-				Flow between pre-chamber and main pipe
-			</description>
-		</method>
-		<method name="get_vrp" qualifiers="static">
-			<return type="float" />
-			<description>
-				Pre-chamber pressure
-			</description>
-		</method>
-		<method name="init_brake" qualifiers="static">
-			<return type="void" />
-			<param index="0" name="_unnamed_arg0" type="float" />
-			<param index="1" name="_unnamed_arg1" type="float" />
-			<param index="2" name="_unnamed_arg2" type="float" />
-			<param index="3" name="_unnamed_arg3" type="float" />
-			<param index="4" name="_unnamed_arg4" type="int" />
-			<param index="5" name="_unnamed_arg5" type="float" />
-			<param index="6" name="_unnamed_arg6" type="float" />
-			<param index="7" name="_unnamed_arg7" type="float" />
-			<param index="8" name="_unnamed_arg8" type="float" />
-			<param index="9" name="_unnamed_arg9" type="int" />
-			<param index="10" name="_unnamed_arg10" type="int" />
-			<param index="11" name="_unnamed_arg11" type="int" />
-			<param index="12" name="_unnamed_arg12" type="int" />
-			<param index="13" name="_unnamed_arg13" type="int" />
-			<description>
-				Initializes brake. What did you expect?
-			</description>
-		</method>
-	</methods>
-</class>
