+++ src/gen/doc_data.gen.cpp
@@ -2,675 +2,549 @@
 
 #include <godot_cpp/godot.hpp>
 
-static const char *_doc_data_hash = "1264283215946149618";
-static const int _doc_data_uncompressed_size = 3470;
-static const int _doc_data_compressed_size = 665;
+static const char *_doc_data_hash = "231522034031474057";
+static const int _doc_data_uncompressed_size = 2534;
+static const int _doc_data_compressed_size = 539;
 static const unsigned char _doc_data_compressed[] = {
 	120,
 	218,
-	229,
-	87,
-	77,
-	111,
-	219,
-	48,
-	12,
-	61,
-	167,
-	191,
-	66,
-	208,
-	189,
-	118,
-	156,
-	244,
-	27,
-	113,
-	139,
-	174,
-	88,
-	129,
-	1,
-	93,
-	49,
-	96,
-	27,
-	182,
-	91,
-	160,
-	72,
-	178,
-	77,
-	212,
+	221,
 	150,
-	52,
-	73,
-	174,
-	227,
-	253,
-	250,
-	209,
-	78,
-	10,
-	244,
-	67,
-	69,
-	82,
-	55,
-	151,
-	98,
-	55,
-	138,
-	210,
-	35,
-	159,
-	158,
-	72,
-	89,
-	158,
-	93,
-	44,
-	171,
-	146,
-	220,
-	75,
-	235,
-	64,
-	171,
-	148,
-	38,
-	209,
-	152,
-	18,
-	169,
-	184,
-	22,
-	160,
-	242,
-	148,
-	254,
-	252,
-	113,
-	189,
-	127,
-	66,
-	201,
-	197,
-	249,
-	222,
-	140,
-	151,
-	204,
-	57,
-	162,
-	88,
-	37,
-	83,
-	122,
-	105,
-	128,
-	18,
-	80,
-	133,
-	180,
-	224,
-	93,
-	74,
-	111,
-	181,
-	144,
-	148,
-	96,
-	28,
-	229,
-	206,
-	150,
-	14,
-	82,
-	90,
-	120,
-	111,
-	206,
-	226,
-	184,
-	105,
-	154,
-	168,
-	153,
-	70,
-	218,
-	230,
-	241,
-	100,
-	60,
-	78,
-	226,
-	223,
-	95,
-	111,
-	190,
-	243,
-	66,
-	86,
-	108,
-	31,
-	148,
-	243,
-	76,
-	241,
-	14,
-	229,
-	224,
-	76,
-	233,
-	91,
-	12,
-	235,
-	12,
-	227,
-	114,
-	53,
-	127,
-	163,
-	57,
-	243,
-	61,
-	159,
-	46,
-	146,
-	195,
-	80,
-	150,
-	53,
-	81,
-	14,
-	190,
-	168,
-	23,
-	181,
-	147,
-	150,
-	107,
-	229,
-	165,
-	242,
-	17,
-	215,
-	85,
-	156,
-	107,
-	161,
-	113,
-	144,
-	131,
-	146,
-	43,
-	59,
-	174,
-	152,
-	243,
-	210,
-	198,
-	66,
-	243,
-	184,
-	39,
-	29,
-	45,
-	157,
-	160,
-	231,
-	123,
-	163,
-	217,
-	194,
-	130,
-	204,
-	230,
-	66,
-	58,
+	193,
 	110,
-	193,
-	116,
-	241,
-	59,
-	103,
-	28,
-	244,
-	62,
-	95,
-	244,
-	108,
-	236,
-	107,
-	175,
-	45,
-	176,
-	210,
-	245,
-	147,
-	143,
-	70,
-	179,
-	85,
-	74,
-	52,
-	134,
-	232,
-	250,
-	201,
-	178,
-	59,
-	249,
-	95,
-	41,
-	59,
-	186,
-	234,
-	183,
-	95,
-	48,
-	37,
-	74,
-	84,
-	134,
-	44,
-	58,
-	5,
-	136,
-	107,
-	49,
-	79,
-	69,
-	26,
-	164,
-	181,
-	246,
-	248,
-	214,
-	72,
-	71,
-	112,
-	21,
-	49,
-	204,
-	226,
-	150,
-	144,
-	134,
-	123,
-	203,
-	185,
-	140,
-	102,
-	136,
-	41,
-	180,
-	232,
-	204,
-	7,
-	123,
-	45,
-	122,
-	46,
-	253,
-	124,
-	193,
-	51,
-	74,
-	254,
-	212,
-	172,
-	132,
-	12,
-	48,
-	114,
-	74,
-	81,
-	68,
-	15,
-	188,
 	219,
-	27,
-	46,
-	183,
-	210,
-	215,
-	86,
-	245,
-	28,
-	82,
-	154,
-	149,
-	154,
+	48,
+	12,
+	134,
+	239,
 	121,
-	74,
-	226,
-	213,
-	212,
-	179,
-	221,
-	188,
-	96,
-	132,
-	142,
-	85,
-	178,
-	215,
-	242,
-	154,
-	221,
-	228,
-	29,
-	245,
-	165,
-	67,
-	120,
-	139,
-	42,
 	10,
-	105,
+	65,
+	247,
+	216,
+	73,
+	118,
+	25,
+	130,
+	56,
+	69,
+	155,
+	53,
+	219,
+	161,
+	45,
+	130,
+	173,
+	27,
+	122,
+	11,
+	20,
 	137,
 	177,
-	210,
-	185,
-	218,
-	202,
-	33,
-	156,
-	236,
-	174,
-	56,
-	93,
-	214,
-	75,
-	40,
-	129,
-	217,
-	150,
-	96,
-	73,
-	222,
-	189,
-	135,
-	19,
-	223,
-	25,
-	167,
-	43,
-	44,
-	112,
-	171,
-	203,
-	119,
-	51,
-	146,
-	98,
-	119,
-	103,
-	247,
-	185,
-	148,
-	28,
-	73,
-	137,
-	22,
-	195,
-	3,
-	95,
-	87,
-	60,
-	95,
-	19,
-	125,
-	7,
-	199,
-	140,
-	15,
-	35,
-	216,
-	183,
-	24,
-	233,
-	10,
-	105,
-	153,
-	82,
-	188,
-	183,
-	86,
-	17,
-	231,
-	181,
-	234,
-	12,
-	49,
-	103,
-	54,
-	71,
-	231,
-	70,
-	96,
-	18,
-	2,
-	38,
-	244,
-	45,
-	229,
-	156,
-	89,
-	224,
-	157,
-	15,
-	197,
-	144,
-	89,
-	6,
-	28,
-	240,
-	106,
-	26,
-	160,
-	67,
-	97,
-	48,
-	91,
-	243,
-	97,
-	181,
-	184,
-	70,
-	242,
-	40,
 	133,
-	174,
-	136,
-	47,
-	240,
-	110,
-	172,
-	141,
+	218,
+	146,
+	33,
+	202,
+	115,
+	188,
+	167,
+	31,
+	101,
+	187,
+	232,
+	214,
+	166,
+	195,
+	208,
+	58,
+	135,
+	77,
+	39,
+	139,
+	146,
+	201,
+	143,
+	191,
 	41,
+	202,
+	139,
+	179,
+	67,
+	145,
+	179,
+	239,
+	224,
+	80,
 	91,
-	130,
-	125,
-	62,
-	164,
-	36,
-	76,
+	147,
+	240,
+	105,
+	52,
+	225,
+	12,
+	140,
+	180,
+	74,
+	155,
+	52,
+	225,
+	95,
+	111,
+	215,
+	227,
+	247,
+	156,
+	157,
+	45,
+	71,
+	11,
+	153,
+	11,
+	68,
+	102,
+	68,
+	1,
+	9,
+	191,
+	22,
+	248,
+	163,
+	49,
+	226,
 	246,
-	81,
+	194,
+	137,
+	123,
+	224,
+	76,
+	155,
+	12,
+	156,
+	246,
+	152,
+	240,
+	27,
+	171,
+	104,
+	78,
+	30,
+	13,
+	206,
+	15,
+	168,
+	19,
+	158,
+	121,
+	95,
+	206,
+	227,
+	184,
+	174,
+	235,
+	168,
+	126,
+	23,
+	89,
+	151,
+	198,
+	179,
+	201,
+	100,
+	26,
+	223,
+	93,
+	95,
+	125,
+	145,
+	25,
+	20,
+	98,
+	172,
+	13,
+	122,
+	97,
 	100,
 	120,
-	2,
-	156,
-	132,
-	128,
-	147,
-	183,
-	232,
-	183,
-	144,
-	190,
-	145,
-	82,
-	117,
-	237,
-	180,
-	207,
 	11,
-	86,
-	45,
-	240,
-	146,
-	236,
-	62,
-	39,
-	21,
-	3,
-	116,
-	130,
-	25,
-	34,
-	230,
-	253,
-	206,
-	110,
-	165,
-	111,
-	143,
-	88,
-	13,
-	108,
-	120,
-	80,
-	208,
-	221,
-	221,
-	253,
-	3,
-	98,
-	59,
-	78,
-	247,
-	26,
-	196,
-	71,
-	60,
-	224,
-	39,
-	192,
-	105,
-	8,
-	56,
-	221,
-	2,
-	120,
-	16,
-	2,
-	30,
-	60,
-	0,
-	65,
-	189,
-	2,
-	59,
-	12,
-	193,
-	14,
-	183,
-	200,
-	119,
-	20,
-	2,
-	30,
-	109,
-	1,
-	60,
-	14,
-	1,
-	143,
-	183,
-	0,
-	158,
-	132,
+	245,
+	220,
+	216,
+	27,
+	10,
 	128,
-	39,
-	91,
-	0,
-	79,
-	67,
-	192,
-	211,
+	165,
+	144,
+	208,
+	173,
+	95,
+	89,
+	41,
+	124,
+	75,
+	22,
+	60,
+	33,
+	185,
+	114,
+	162,
+	142,
+	82,
+	237,
+	179,
+	106,
+	87,
+	33,
+	56,
+	105,
 	141,
+	7,
+	227,
+	35,
+	105,
+	139,
+	56,
+	181,
+	202,
 	210,
 	36,
-	193,
-	178,
-	73,
-	198,
-	155,
-	129,
-	225,
-	178,
-	73,
-	54,
-	3,
-	131,
-	101,
-	147,
-	76,
-	54,
-	3,
-	131,
-	101,
-	147,
-	76,
-	67,
-	192,
-	23,
-	205,
-	250,
-	5,
-	59,
-	13,
-	95,
-	149,
-	240,
-	23,
-	95,
-	164,
-	125,
-	195,
-	69,
-	228,
-	87,
-	193,
-	60,
-	17,
-	32,
-	72,
-	171,
-	107,
-	34,
-	151,
+	213,
 	6,
-	191,
+	186,
 	231,
-	23,
-	155,
-	250,
-	247,
+	184,
+	16,
+	232,
 	193,
-	124,
-	252,
-	215,
-	240,
-	15,
-	222,
-	223,
-	90,
+	197,
+	202,
+	202,
+	184,
+	197,
+	143,
+	14,
+	168,
+	248,
+	114,
+	196,
+	104,
+	44,
+	118,
+	78,
+	195,
+	126,
+	171,
+	0,
+	165,
+	211,
+	101,
+	8,
+	209,
+	219,
+	227,
+	151,
+	22,
+	158,
+	89,
+	194,
+	88,
+	181,
+	170,
+	100,
+	194,
+	168,
+	156,
+	4,
 	99,
+	187,
+	32,
+	8,
+	195,
+	134,
+	194,
+	22,
+	172,
+	38,
+	202,
+	222,
+	226,
+	155,
+	18,
+	144,
+	209,
+	46,
+	86,
+	10,
+	71,
+	25,
+	18,
+	21,
+	246,
+	225,
+	158,
+	7,
+	242,
+	149,
+	183,
+	78,
+	139,
+	28,
+	31,
+	136,
+	158,
+	26,
+	232,
+	253,
+	204,
+	42,
+	124,
+	164,
+	232,
+	45,
+	253,
+	215,
+	249,
+	8,
+	254,
+	98,
+	181,
+	230,
+	236,
+	113,
+	189,
+	221,
+	227,
+	192,
+	87,
+	206,
+	180,
+	40,
+	9,
+	223,
+	231,
+	86,
+	120,
+	206,
+	226,
+	39,
+	123,
+	142,
+	230,
+	248,
+	2,
+	103,
+	103,
+	238,
+	34,
+	255,
+	17,
+	101,
+	51,
+	44,
+	74,
+	24,
+	109,
+	225,
+	49,
+	217,
+	144,
+	232,
+	10,
+	28,
+	43,
+	29,
+	32,
+	86,
+	14,
+	134,
+	66,
+	254,
+	124,
+	2,
+	228,
+	243,
+	234,
+	160,
+	115,
+	45,
+	92,
+	195,
+	168,
+	222,
+	239,
+	135,
+	70,
+	94,
+	157,
+	2,
+	121,
+	69,
+	135,
+	203,
+	217,
+	252,
+	36,
+	192,
+	151,
+	31,
+	78,
+	82,
+	24,
+	151,
+	57,
+	72,
+	98,
+	86,
+	212,
+	157,
+	10,
+	45,
+	251,
+	211,
+	39,
+	251,
+	60,
+	6,
+	78,
+	97,
+	189,
+	122,
+	21,
+	127,
+	219,
+	1,
+	88,
+	40,
+	220,
+	67,
+	194,
+	169,
+	219,
+	118,
+	254,
+	182,
+	149,
+	9,
+	15,
+	106,
+	43,
+	92,
+	74,
+	198,
+	191,
+	247,
+	48,
+	61,
+	230,
+	97,
+	202,
+	223,
+	124,
+	184,
+	246,
+	78,
+	203,
+	176,
+	78,
+	234,
+	193,
+	126,
+	175,
+	165,
+	166,
+	54,
+	59,
+	144,
+	112,
+	159,
+	54,
+	235,
+	220,
+	214,
+	255,
+	167,
+	120,
+	33,
+	51,
+	210,
+	206,
+	22,
+	204,
+	103,
+	116,
+	19,
+	84,
+	101,
+	153,
+	55,
+	140,
+	218,
+	212,
+	80,
+	69,
+	183,
+	89,
+	255,
+	179,
+	186,
+	253,
+	230,
+	97,
+	118,
+	204,
+	195,
+	236,
+	205,
+	202,
+	239,
+	192,
+	215,
+	0,
+	38,
+	28,
+	245,
+	177,
+	204,
+	68,
+	177,
+	163,
+	219,
+	33,
+	92,
+	187,
+	133,
+	208,
+	100,
+	212,
+	229,
+	80,
+	159,
+	225,
+	219,
+	41,
+	250,
+	237,
+	230,
+	23,
+	232,
+	215,
+	247,
+	170,
+	135,
+	25,
+	253,
+	31,
+	44,
+	186,
+	223,
+	158,
+	229,
+	232,
+	39,
+	173,
+	125,
+	198,
+	199,
 };
 
 static godot::internal::DocDataRegistration _doc_data_registration(_doc_data_hash, _doc_data_uncompressed_size, _doc_data_compressed_size, _doc_data_compressed);
