+++ src/register_types.cpp
@@ -4,32 +4,33 @@
 #include <godot_cpp/core/defs.hpp>
 #include <godot_cpp/godot.hpp>
 
-#include "maszyna.h"
+#include "MaszynaTBrake.h"
 
 using namespace godot;
 
-void initialize_example_module(ModuleInitializationLevel p_level) {
-	if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {
-		return;
-	}
+void initialize_libmaszyna_module(ModuleInitializationLevel p_level) {
+    if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {
+        return;
+    }
 
-	GDREGISTER_CLASS(Maszyna::Api);
-	GDREGISTER_CLASS(Maszyna::Brakes::Brake);
+    GDREGISTER_CLASS(MaszynaTBrake);
 }
 
-void uninitialize_example_module(ModuleInitializationLevel p_level) {
-	if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {}
+void uninitialize_libmaszyna_module(ModuleInitializationLevel p_level) {
+    if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {
+        return;
+    }
 }
 
 extern "C" {
-	// Initialization.
-	GDExtensionBool GDE_EXPORT example_library_init(GDExtensionInterfaceGetProcAddress p_get_proc_address, const GDExtensionClassLibraryPtr p_library, GDExtensionInitialization *r_initialization) {
-		const GDExtensionBinding::InitObject init_obj(p_get_proc_address, p_library, r_initialization);
+    // Initialization.
+    GDExtensionBool GDE_EXPORT libmaszyna_library_init(GDExtensionInterfaceGetProcAddress p_get_proc_address, const GDExtensionClassLibraryPtr p_library, GDExtensionInitialization *r_initialization) {
+        const GDExtensionBinding::InitObject init_obj(p_get_proc_address, p_library, r_initialization);
 
-		init_obj.register_initializer(initialize_example_module);
-		init_obj.register_terminator(uninitialize_example_module);
-		init_obj.set_minimum_library_initialization_level(MODULE_INITIALIZATION_LEVEL_SCENE);
+        init_obj.register_initializer(initialize_libmaszyna_module);
+        init_obj.register_terminator(uninitialize_libmaszyna_module);
+        init_obj.set_minimum_library_initialization_level(MODULE_INITIALIZATION_LEVEL_SCENE);
 
-		return init_obj.init();
-	}
+        return init_obj.init();
+    }
 }
