+++ src/register_types.h
@@ -5,7 +5,7 @@
 
 using namespace godot;
 
-void initialize_example_module(ModuleInitializationLevel p_level);
-void uninitialize_example_module(ModuleInitializationLevel p_level);
+void initialize_libmaszyna_module(ModuleInitializationLevel p_level);
+void uninitialize_libmaszyna_module(ModuleInitializationLevel p_level);
 
 #endif // GDEXAMPLE_REGISTER_TYPES_H
